requirements.txt
bs4
vnstock
pandas
google-cloud-storage
pyspark
google.cloud.pubsub